refactor(frontend): update search results for themes
