+++ packages/code-du-travail-frontend/pages/fiche-ministere-travail.js
@@ -4,7 +4,7 @@ import getConfig from "next/config";
 import fetch from "isomorphic-unfetch";
 import { ExternalLink } from "react-feather";
 import Answer from "../src/common/Answer";
-import DossierIcon from "../src/icons/DossierIcon";
+import ReponseIcon from "../src/icons/ReponseIcon";
 import { PageLayout } from "../src/layout/PageLayout";
 
 const {
@@ -47,7 +47,7 @@ class Fiche extends React.Component {
           emptyMessage="Cette fiche n'a pas été trouvée"
           html={data._source.html}
           footer={footer}
-          icon={DossierIcon}
+          icon={ReponseIcon}
           date={data._source.date}
           sourceType="Fiche ministère du travail"
         />
